rt-dag: suppress signedness compiler warning

+++ tools/rt-dag/rt-dag/include/input_yaml.h
@@ -41,7 +41,7 @@ public:
         //cout << "Reading name: " << inputs["dag_name"] << endl;
         dag_name = inputs["dag_name"].as<string>();
         const int ntasks = get_n_tasks();
-        if (ntasks >= sizeof(tasks_name)/sizeof(tasks_name[0])) {
+        if (ntasks >= (int) (sizeof(tasks_name)/sizeof(tasks_name[0]))) {
             cerr << "Max # of tasks exceeded in input YAML" << endl;
             exit(EXIT_FAILURE);
         }
